requirements.txt
streamlit
PyPDF2
python-dotenv
langchain-openai
langchain-google-genai
faiss-cpu